Cleanup and package ui
